requirements.txt
langchain
langchain-google-genai
cachetools
python-dotenv